make it easier to run SOI setter locally and fix package name
