+++ handlers/soft-opt-in-consent-setter/src/main/scala/com/gu/soft_opt_in_consent_setter/Handler.scala
@@ -26,7 +26,7 @@ object Handler extends LazyLogging {
 
   def handleRequest(): Unit = {
     (for {
-      config <- SoftOptInConfig()
+      config <- SoftOptInConfig(sys.env.get("Stage"), sys.env.get("sfApiVersion"))
     } yield for {
       sfConnector <- SalesforceConnector(config.sfConfig, config.sfApiVersion)
