+++ handlers/soft-opt-in-consent-setter/src/main/scala/com/gu/soft_opt_in_consent_setter/models/Secrets.scala
@@ -1,8 +1,11 @@
 package com.gu.soft_opt_in_consent_setter.models
 
+import com.gu.soft_opt_in_consent_setter.AwsCredentialsBuilder
+import software.amazon.awssdk.auth.credentials.AwsCredentialsProviderChain
 import software.amazon.awssdk.services.secretsmanager._
 import software.amazon.awssdk.services.secretsmanager.model.GetSecretValueRequest
 import upickle.default._
+
 import scala.util.Try
 
 /*
@@ -42,12 +45,16 @@ case class MobilePurchasesAPIUserGetSubscriptionsSecrets(mpapiUrl: String, mpapi
 
 object Secrets {
 
+  def apply(stg: String): Either[SoftOptInError, Secrets] = {
+    AwsCredentialsBuilder.buildCredentials
+      .map(SecretsManagerClient.builder().credentialsProvider(_).build())
+      .map(new Secrets(stg, _))
+  }
+
   implicit val reader1: Reader[SalesforceConnectedAppSecrets] = macroRW
   implicit val reader2: Reader[SalesforceUserSecrets] = macroRW
   implicit val reader3: Reader[IdentitySoftOptInConsentAPISecrets] = macroRW
   implicit val reader4: Reader[MobilePurchasesAPIUserGetSubscriptionsSecrets] = macroRW
-
-  private lazy val secretsClient = SecretsManagerClient.create()
 
   val stageToSalesforceStage: Map[String, String] = Map(
     "CODE" -> "CODE",
@@ -106,58 +113,62 @@ object Secrets {
 
   def salesforceStageSecretsId(salesforceStage: String, appName: String) =
     s"${salesforceStage}/Salesforce/ConnectedApp/${appName}"
+
   def salesforceUserSecretsId(salesforceStage: String, salesforceUsername: String) =
     s"${salesforceStage}/Salesforce/User/${salesforceUsername}"
+
   def identitySoftOptInConsentAPISecretsSecretsId(identityStage: String) =
     s"${identityStage}/Identity/SoftOptInConsentAPI"
+
   def mobilePurchasesAPIUserGetSubscriptionsSecretsSecretsId(mpapiStage: String) =
     s"${mpapiStage}/MobilePurchasesAPI/User/GetSubscriptions"
+
+}
+class Secrets(stg: String, secretsClient: SecretsManagerClient) {
+
+  import Secrets._
 
   def getJSONString(secretId: String): String = {
     secretsClient.getSecretValue(GetSecretValueRequest.builder().secretId(secretId).build()).secretString()
   }
 
-  lazy val stage: Option[String] = sys.env.get("Stage")
-
-  def getSalesforceConnectedAppSecrets: Option[SalesforceConnectedAppSecrets] = {
+  def getSalesforceConnectedAppSecrets: Either[String, SalesforceConnectedAppSecrets] = {
     for {
-      stg <- stage
-      salesforceStage <- stageToSalesforceStage.get(stg)
-      appName <- stageToAppname.get(stg)
+      salesforceStage <- stageToSalesforceStage.get(stg).toRight("salesforceStage is missing")
+      appName <- stageToAppname.get(stg).toRight("appName is missing")
       secretId = salesforceStageSecretsId(salesforceStage, appName)
       secretJsonString = getJSONString(secretId)
-      secrets <- Try(read[SalesforceConnectedAppSecrets](secretJsonString)).toOption
+      secrets <- Try(read[SalesforceConnectedAppSecrets](secretJsonString)).toEither.left.map(_.toString)
     } yield secrets
   }
 
-  def getSalesforceUserSecrets: Option[SalesforceUserSecrets] = {
+  def getSalesforceUserSecrets: Either[String, SalesforceUserSecrets] = {
     for {
-      stg <- stage
-      salesforceStage <- stageToSalesforceStage.get(stg)
-      stageToSalesforceUsername <- stageToSalesforceUsername.get(stg)
+      salesforceStage <- stageToSalesforceStage.get(stg).toRight("salesforceStage is missing")
+      stageToSalesforceUsername <- stageToSalesforceUsername.get(stg).toRight("stageToSalesforceUsername is missing")
       secretId = salesforceUserSecretsId(salesforceStage, stageToSalesforceUsername)
       secretJsonString = getJSONString(secretId)
-      secrets <- Try(read[SalesforceUserSecrets](secretJsonString)).toOption
+      secrets <- Try(read[SalesforceUserSecrets](secretJsonString)).toEither.left.map(_.toString)
     } yield secrets
   }
 
-  def getIdentitySoftOptInConsentAPISecrets: Option[IdentitySoftOptInConsentAPISecrets] = {
+  def getIdentitySoftOptInConsentAPISecrets: Either[String, IdentitySoftOptInConsentAPISecrets] = {
     for {
-      stg <- stage
-      identityStage <- stageToIdentityStage.get(stg)
+      identityStage <- stageToIdentityStage.get(stg).toRight("identityStage is missing")
       secretId = identitySoftOptInConsentAPISecretsSecretsId(identityStage)
       secretJsonString = getJSONString(secretId)
-      secrets <- Try(read[IdentitySoftOptInConsentAPISecrets](secretJsonString)).toOption
+      secrets <- Try(read[IdentitySoftOptInConsentAPISecrets](secretJsonString)).toEither.left.map(_.toString)
     } yield secrets
   }
 
-  def getMobilePurchasesAPIUserGetSubscriptionsSecrets: Option[MobilePurchasesAPIUserGetSubscriptionsSecrets] = {
+  def getMobilePurchasesAPIUserGetSubscriptionsSecrets
+      : Either[String, MobilePurchasesAPIUserGetSubscriptionsSecrets] = {
     for {
-      stg <- stage
-      mpapiStage <- stageMpapiStage.get(stg)
+      mpapiStage <- stageMpapiStage.get(stg).toRight("mpapiStage is missing")
       secretId = mobilePurchasesAPIUserGetSubscriptionsSecretsSecretsId(mpapiStage)
       secretJsonString = getJSONString(secretId)
-      secrets <- Try(read[MobilePurchasesAPIUserGetSubscriptionsSecrets](secretJsonString)).toOption
+      secrets <- Try(read[MobilePurchasesAPIUserGetSubscriptionsSecrets](secretJsonString)).toEither.left
+        .map(_.toString)
     } yield secrets
   }
 }
